Set FCS included flag in Radiotap header
This patch addresses Bug 875 by correcting a bug in the YansWifiHelper
support for PCAP tracing with Radiotap data link format. Specifically,
in the previous implementation the "Frame includes FCS" flag is not
set, despite the FCS being present in the captured frames.

This changeset requires a corresponding change to the
wifi-wired-bridging reference traces (this test, which is run as part
of the regression test set, appears to be the only one which uses
Radiotap capture).